cipp_exp: Change CIPP interleave interval from 10s to 15s
We increase this slightly to make it double the operating interval
of DAMON. We want it to be double so that at least the second half
of CIPP's bandwidth measurements are with the previous ratio applied.
We have to worry about this because CIPP and DAMON do not directly
communicate.